backend/requirements.txt
fastapi
uvicorn
pydantic
python-jose[cryptography]
passlib[bcrypt]
asyncpg
python-multipart
